Add rounded parameter to Init function and update price rounding in depth analyzer

+++ utils/round.go
@@ -2,7 +2,7 @@ package utils
 
 import "math"
 
-func RoundToDecimalPlace(x float64, exp float64) float64 {
+func RoundToDecimalPlace(x float64, exp int) float64 {
 	if exp >= 0 {
 		pow := math.Pow(10, float64(exp))
 		return math.Round(x*pow) / pow
